add isolation level around polling requests and submitting jobs.
SERIALIZABLE is the highest isolation level and may introduce a performance overhead (which is small since job submission is asynchronous). But this is the only way to ensure that multiple servers do not submit two jobs for the same request and put the system in an inconsistent state.

Consistency in horizontal scaling is by far more important than this performance overhead.